Project update
Only spend few days working on this.

so far in this new build.

you can connect to server. ( ATM it best to load unity they start the server) - I am looking in to this

You can make chars but only war flab have been loaded in.

all Prefabs need updating or adding and Animations need added only war so far for testing.

Loading maps is next on the to do list for the new server.